requirements-docker.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-socks==0.7.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
docker==7.0.0
dropbox==11.36.2
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
MarkupSafe==2.1.5
matrix-nio==0.23.0
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
ply==3.11
pycparser==2.21
pycryptodome==3.20.0
python-cryptography-fernet-wrapper==1.0.4
python-dateutil==2.8.2
python-socks==2.4.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
setuptools==69.0.3
simplematrixbotlib==2.10.3
six==1.16.0
stone==3.3.1
toml==0.10.2
tzdata==2023.4
unpaddedbase64==2.1.0
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
